pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "habit-tracking-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Макаров Артём",email = "makarovartem.09@mail.ru"}
]
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pytelegrambotapi (>=4.33.0,<5.0.0)",
    "fastapi (>=0.136.0,<0.137.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "aiohttp (>=3.13.5,<4.0.0)",
    "pydentic (>=0.0.1.dev3,<0.0.2)",
    "black (>=26.3.1,<27.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "vk-api (>=11.10.0,<12.0.0)",
    "celery[redis] (>=5.6.3,<6.0.0)",
    "psycopg2 (>=2.9.12,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
